add viewmodel databinding and sqlite database

+++ app/src/main/java/com/maxrzhe/contactsapp/model/Contact.kt
@@ -4,14 +4,14 @@ import android.os.Parcel
 import android.os.Parcelable
 
 data class Contact(
-    val id: Int,
+    val id: Long,
     val name: String?,
     val phone: String?,
     val email: String?,
     val image: String?
 ): Parcelable {
     constructor(parcel: Parcel) : this(
-        parcel.readInt(),
+        parcel.readLong(),
         parcel.readString(),
         parcel.readString(),
         parcel.readString(),
@@ -19,7 +19,7 @@ data class Contact(
     )
 
     override fun writeToParcel(parcel: Parcel, flags: Int) {
-        parcel.writeInt(id)
+        parcel.writeLong(id)
         parcel.writeString(name)
         parcel.writeString(phone)
         parcel.writeString(email)
